Fix up gateway annotation with binding logic
